Optimize Docker Build and Adjust Redis Config
Added the `BUILD_ENV` variable to conditionally skip execution of certain parts of the code that are not relevant during the build, such as the Django signal to populate cache.

Additionally, adjusted the Redis cache location configuration in `common.py`, changing it to a base path without specifying a database number.